@@ -1,5 +1,5 @@
-//go:build !prysmAt1.20
-// +build !prysmAt1.20
+//go:build !go1.20
+// +build !go1.20
 
 package bytesutil
 
